apply modern-cpp-template to the lib
applied this
https://github.com/filipdutescu/modern-cpp-template

because current cmake sucks.

Please, do not lie about that this is can be not header-only lib. It IS
header only lib. Maybe it does precompiling headers ( including header
to cpp and nothing to do), but in case of the lib, where there are a lot
of templates, it's nearly useless.

Also removed test that was checking ability of linking the library,
because it's pointless, if there's more useful test which needs the lib.
If those fails, we can see it in compilation failing at preprocessing
stage.

Edited README with adding explanation how to using the lib.

I didn't changed the code of the lib, except in fmtlog.h it adds
fmtlog-inl.h anytime, because it's actually a header-only lib and let's
not say it can be not header only just by adding one header to a .cc file.